+++ src/test/scala/com/atomist/rug/kind/RugIsExtensible.scala
@@ -16,29 +16,41 @@ import _root_.scala.util.parsing.combinator.RegexParsers
 
 class PandaRugLanguageExtension extends TypeUnderFile {
 
+  import com.atomist.rug.kind.grammar.ParsedNode
   import com.atomist.source.FileArtifact
-  import com.atomist.tree.content.text.PositionedTreeNode
 
   override val name: String = "Panda" // To trigger parsing, get to a file in the path expression, then /Panda()
 
   override def isOfType(f: FileArtifact): Boolean = f.name.endsWith(".panda")
 
-  override def fileToRawNode(f: FileArtifact): Option[PositionedTreeNode] =
+  override def fileToRawNode(f: FileArtifact): Option[ParsedNode] =
     Some(PandaParser.parse(f.content))
 
   override def description: String = "A language extension for my imaginary language"
 
+  override def preprocess(originalContent: String): String =
+    originalContent.replaceAll("(?m)^(\\S)", "\\$$1")
+
+
+  override def postprocess(preprocessedContent: String): String =
+     preprocessedContent.replaceAll("(?m)^\\$", "")
+
 }
 
 object Panda {
+
   val PandaText =
     """panda { panda panda } panda
-      |(panda panda) panda
+      | (panda panda) panda
+      |
+      |panda panda
     """.stripMargin
 
   val KawaiiPandas =
-    """panda { kawaii kawaii } panda
-      |(panda panda) panda
+    """bear { kawaii kawaii } panda
+      | (panda panda) panda
+      |
+      |bear panda
     """.stripMargin
 
   val PandaFilename = "my.panda"
@@ -51,55 +63,12 @@ object Panda {
 
 object PandaParser extends RegexParsers {
 
-  import com.atomist.tree.content.text.{OffsetInputPosition, PositionedTreeNode}
+  import com.atomist.rug.kind.grammar.{ParsedNode, SimpleParsedNode}
 
-  trait MinimalPositionedTreeNode extends PositionedTreeNode {
-
-    import com.atomist.tree.TreeNode
-
-    def name: String
-
-    def valueOption: Option[String]
-
-    def start: Int
-
-    def end: Int
-
-    override def nodeName = name
-
-    override def startPosition = OffsetInputPosition(start)
-
-    override def endPosition = OffsetInputPosition(end)
-
-    override def value = valueOption.getOrElse("")
-
-    override def childNodeNames: Set[String] = ???
-
-    override def childNodeTypes: Set[String] = ???
-
-    override def childrenNamed(key: String): Seq[TreeNode] = childNodes.filter(_.nodeName == key)
-
-  }
-
-  case class PositionedSyntaxNode(name: String,
-                                  override val childNodes: Seq[PositionedSyntaxNode],
-                                  override val valueOption: Option[String],
-                                  start: Int,
-                                  end: Int) extends MinimalPositionedTreeNode
-
+  type PositionedSyntaxNode = ParsedNode
 
   case class SyntaxNode(name: String,
-                        childNodes: Seq[PositionedSyntaxNode],
-                        valueOption: Option[String])
-
-  object SyntaxNode {
-
-    def leaf(name: String)(value: String): SyntaxNode =
-      SyntaxNode(name, Seq(), Some(value))
-
-    def parent(name: String, children: Seq[PositionedSyntaxNode]): SyntaxNode =
-      SyntaxNode(name, children, None)
-  }
+                        childNodes: Seq[ParsedNode] = Seq())
 
   def positionedNode(inner: Parser[SyntaxNode]) = new Parser[PositionedSyntaxNode] {
     override def apply(in: Input): ParseResult[PositionedSyntaxNode] = {
@@ -109,23 +78,25 @@ object PandaParser extends RegexParsers {
         case Failure(msg, next) => Failure(msg, next)
         case Success(result, next) =>
           val end = next.offset
-          Success(PositionedSyntaxNode(result.name, result.childNodes, result.valueOption,
-            start, end), next)
+          Success(SimpleParsedNode(result.name, start, end, result.childNodes), next)
       }
     }
   }
 
+  def line: Parser[PositionedSyntaxNode] =
+    positionedNode("$" ~> rep(word | curly | paren) ^^ { parts => SyntaxNode("line", parts) })
+
   def curly: Parser[PositionedSyntaxNode] =
-    positionedNode("{" ~> rep(word) <~ "}" ^^ { n => SyntaxNode.parent("curly", n) })
+    positionedNode("{" ~> rep(word) <~ "}" ^^ { n => SyntaxNode("curly", n) })
 
   def paren: Parser[PositionedSyntaxNode] =
-    positionedNode("(" ~> rep(word) <~ ")" ^^ { n => SyntaxNode.parent("paren", n) })
+    positionedNode("(" ~> rep(word) <~ ")" ^^ { n => SyntaxNode("paren", n) })
 
   def word: Parser[PositionedSyntaxNode] =
-    positionedNode("[a-zA-Z]+".r ^^ SyntaxNode.leaf("word"))
+    positionedNode("[a-zA-Z]+".r ^^ { _ => SyntaxNode("word") })
 
   def pandaParser: Parser[PositionedSyntaxNode] =
-    positionedNode(rep(word | curly | paren) ^^ { parts => SyntaxNode.parent("panda", parts) })
+    positionedNode(rep(line) ^^ { parts => SyntaxNode("panda", parts) })
 
   def parse(content: String): PositionedSyntaxNode = {
     parseAll(pandaParser, content) match {
@@ -143,9 +114,12 @@ class RugIsExtensible extends FunSpec {
 
       // TODO: get this Panda type declared in the manifest thinger somehow
 
-      val changedProject = runTypeScriptProgram(Panda.PandaProject,
+      val changedProject = runTypeScriptProgram(
+        Panda.PandaProject,
         "com/atomist/rug/kind/PandaParsingTypeScriptTest.ts",
-        Map("changePandasInCurliesTo" -> "kawaii"))
+        Map(
+          "changePandasInCurliesTo" -> "kawaii",
+          "changeFirstPandaInEachLineTo" -> "bear"))
 
       assert(changedProject.findFile(Panda.PandaFilename).get.content == Panda.KawaiiPandas)